+++ src/test/java/org/dbos/apiary/PostgresTests.java
@@ -50,11 +50,7 @@ public class PostgresTests {
     public void resetTables() {
         try {
             PostgresConnection conn = new PostgresConnection("localhost", ApiaryConfig.postgresPort, "postgres", "postgres", "dbos");
-            conn.dropTable("FuncInvocations");
-            conn.dropTable("KVTable");
-            conn.createTable("KVTable", "KVKey integer PRIMARY KEY NOT NULL, KVValue integer NOT NULL");
-            conn.dropTable("KVTableTwo");
-            conn.createTable("KVTableTwo", "KVKeyTwo integer PRIMARY KEY NOT NULL, KVValueTwo integer NOT NULL");
+            conn.dropTable(ProvenanceBuffer.PROV_FuncInvocations);
             conn.dropTable("RetwisPosts");
             conn.createTable("RetwisPosts", "UserID integer NOT NULL, PostID integer NOT NULL, Timestamp integer NOT NULL, Post varchar(1000) NOT NULL");
             conn.dropTable("RetwisFollowees");
@@ -266,419 +262,5 @@ public class PostgresTests {
         assertTrue(Arrays.asList(timeline).contains("hello0"));
         assertTrue(Arrays.asList(timeline).contains("hello1"));
         assertTrue(Arrays.asList(timeline).contains("hello2"));
-    }
-
-    @Test
-    public void testPostgresProvenance() throws InvalidProtocolBufferException, SQLException, InterruptedException {
-        logger.info("testPostgresProvenance");
-
-        PostgresConnection conn = new PostgresConnection("localhost", ApiaryConfig.postgresPort, "postgres", "postgres", "dbos");
-
-        apiaryWorker = new ApiaryWorker(new ApiaryNaiveScheduler(), 4, ApiaryConfig.postgres, ApiaryConfig.provenanceDefaultAddress);
-        apiaryWorker.registerConnection(ApiaryConfig.postgres, conn);
-        apiaryWorker.registerFunction("PostgresProvenanceBasic", ApiaryConfig.postgres, PostgresProvenanceBasic::new);
-        apiaryWorker.startServing();
-
-        ProvenanceBuffer provBuff = apiaryWorker.workerContext.provBuff;
-        assert(provBuff != null);
-
-        // Wait a bit so previous provenance capture data would be flushed out.
-        Thread.sleep(ProvenanceBuffer.exportInterval * 4);
-        Connection provConn = provBuff.conn.get();
-        Statement stmt = provConn.createStatement();
-        String[] tables = {"FUNCINVOCATIONS", "KVTableEvents"};
-        for (String table : tables) {
-            stmt.execute(String.format("TRUNCATE TABLE %s;", table));
-        }
-
-        ApiaryWorkerClient client = new ApiaryWorkerClient("localhost");
-
-        int res;
-        int key = 10, value = 100;
-        res = client.executeFunction("PostgresProvenanceBasic", key, value).getInt();
-        assertEquals(101, res);
-
-        Thread.sleep(ProvenanceBuffer.exportInterval * 2);
-
-        // Check provenance tables.
-        // Check function invocation table.
-        String table = ProvenanceBuffer.PROV_FuncInvocations;
-        ResultSet rs = stmt.executeQuery(String.format("SELECT * FROM %s ORDER BY %s DESC;", table, ProvenanceBuffer.PROV_APIARY_TIMESTAMP));
-        rs.next();
-        long txid1 = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        long resExecId = rs.getLong(ProvenanceBuffer.PROV_EXECUTIONID);
-        String resService = rs.getString(ProvenanceBuffer.PROV_SERVICE);
-        String resFuncName = rs.getString(ProvenanceBuffer.PROV_PROCEDURENAME);
-        assertEquals("DefaultService", resService);
-        assertEquals(PostgresProvenanceBasic.class.getName(), resFuncName);
-
-        rs.next();
-        long txid2 = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        resExecId = rs.getLong(ProvenanceBuffer.PROV_EXECUTIONID);
-        resService = rs.getString(ProvenanceBuffer.PROV_SERVICE);
-        resFuncName = rs.getString(ProvenanceBuffer.PROV_PROCEDURENAME);
-        assertEquals("DefaultService", resService);
-        assertEquals(PostgresProvenanceBasic.class.getName(), resFuncName);
-
-        // Inner transaction should have the same transaction ID.
-        assertEquals(txid1, txid2);
-
-        // Check KVTable.
-        table = "KVTableEvents";
-        int expectedSeqNum = 0;
-        rs = stmt.executeQuery(String.format("SELECT * FROM %s ORDER BY %s;", table, ProvenanceBuffer.PROV_APIARY_TIMESTAMP));
-        rs.next();
-
-        // Should be a lookup for key=1. But with NULL data.
-        long resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        int resExportOp = rs.getInt(ProvenanceBuffer.PROV_APIARY_OPERATION_TYPE);
-        int resKey = rs.getInt("KVKey");
-        assertTrue(rs.wasNull());
-        int resValue = rs.getInt("KVValue");
-        assertTrue(rs.wasNull());
-        int resSeqNum = rs.getInt(ProvenanceBuffer.PROV_QUERY_SEQNUM);
-        assertEquals(expectedSeqNum, resSeqNum);
-        expectedSeqNum += 1;
-        assertEquals(txid2, resTxid);
-        assertEquals(ProvenanceBuffer.ExportOperation.READ.getValue(), resExportOp);
-
-        rs.next();
-        // Should be an insert for key=1.
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        resExportOp = rs.getInt(ProvenanceBuffer.PROV_APIARY_OPERATION_TYPE);
-        resKey = rs.getInt("KVKey");
-        resValue = rs.getInt("KVValue");
-        resSeqNum = rs.getInt(ProvenanceBuffer.PROV_QUERY_SEQNUM);
-        assertEquals(expectedSeqNum, resSeqNum);
-        expectedSeqNum += 1;
-        assertEquals(txid2, resTxid);
-        assertEquals(ProvenanceBuffer.ExportOperation.INSERT.getValue(), resExportOp);
-        assertEquals(1, resKey);
-        assertEquals(value, resValue);
-
-        // Should be an insert for the key value.
-        rs.next();
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        assertEquals(txid1, resTxid);
-        resExportOp = rs.getInt(ProvenanceBuffer.PROV_APIARY_OPERATION_TYPE);
-        resKey = rs.getInt("KVKey");
-        resValue = rs.getInt("KVValue");
-        resSeqNum = rs.getInt(ProvenanceBuffer.PROV_QUERY_SEQNUM);
-        assertEquals(expectedSeqNum, resSeqNum);
-        expectedSeqNum += 1;
-        assertEquals(ProvenanceBuffer.ExportOperation.INSERT.getValue(), resExportOp);
-        assertEquals(key, resKey);
-        assertEquals(value, resValue);
-
-        // Should be a read.
-        rs.next();
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        assertEquals(txid1, resTxid);
-        resExportOp = rs.getInt(ProvenanceBuffer.PROV_APIARY_OPERATION_TYPE);
-        resKey = rs.getInt("KVKey");
-        resValue = rs.getInt("KVValue");
-        resSeqNum = rs.getInt(ProvenanceBuffer.PROV_QUERY_SEQNUM);
-        assertEquals(expectedSeqNum, resSeqNum);
-        expectedSeqNum += 1;
-        assertEquals(ProvenanceBuffer.ExportOperation.READ.getValue(), resExportOp);
-        assertEquals(key, resKey);
-        assertEquals(100, resValue);
-
-        // Should be an update.
-        rs.next();
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        assertEquals(txid1, resTxid);
-        resExportOp = rs.getInt(ProvenanceBuffer.PROV_APIARY_OPERATION_TYPE);
-        resKey = rs.getInt("KVKey");
-        resValue = rs.getInt("KVValue");
-        resSeqNum = rs.getInt(ProvenanceBuffer.PROV_QUERY_SEQNUM);
-        assertEquals(expectedSeqNum, resSeqNum);
-        expectedSeqNum += 1;
-        assertEquals(ProvenanceBuffer.ExportOperation.UPDATE.getValue(), resExportOp);
-        assertEquals(key, resKey);
-        assertEquals(value+1, resValue);
-
-        // Should be a read again.
-        rs.next();
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        assertEquals(txid1, resTxid);
-        resExportOp = rs.getInt(ProvenanceBuffer.PROV_APIARY_OPERATION_TYPE);
-        resKey = rs.getInt("KVKey");
-        resValue = rs.getInt("KVValue");
-        resSeqNum = rs.getInt(ProvenanceBuffer.PROV_QUERY_SEQNUM);
-        assertEquals(expectedSeqNum, resSeqNum);
-        expectedSeqNum += 1;
-        assertEquals(ProvenanceBuffer.ExportOperation.READ.getValue(), resExportOp);
-        assertEquals(key, resKey);
-        assertEquals(101, resValue);
-
-        // Should be a delete.
-        rs.next();
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        assertEquals(txid1, resTxid);
-        resExportOp = rs.getInt(ProvenanceBuffer.PROV_APIARY_OPERATION_TYPE);
-        resKey = rs.getInt("KVKey");
-        resValue = rs.getInt("KVValue");
-        resSeqNum = rs.getInt(ProvenanceBuffer.PROV_QUERY_SEQNUM);
-        assertEquals(expectedSeqNum, resSeqNum);
-        assertEquals(ProvenanceBuffer.ExportOperation.DELETE.getValue(), resExportOp);
-        assertEquals(key, resKey);
-        assertEquals(value+1, resValue);
-
-        // Check Query Metadata table.
-        table = "kvtable";
-        String metatable = ProvenanceBuffer.PROV_QueryMetadata;
-        String projection = "kvvalue,kvkey";
-        expectedSeqNum = 0;
-        rs = stmt.executeQuery(String.format("SELECT * FROM %s WHERE %s != 'apiarymetadata' ORDER BY %s;", metatable, ProvenanceBuffer.PROV_QUERY_TABLENAMES, ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID));
-        rs.next();
-
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        resSeqNum = rs.getInt(ProvenanceBuffer.PROV_QUERY_SEQNUM);
-        String resTableName = rs.getString(ProvenanceBuffer.PROV_QUERY_TABLENAMES);
-        String resProjection = rs.getString(ProvenanceBuffer.PROV_QUERY_PROJECTION);
-        assertEquals(expectedSeqNum, resSeqNum);
-        expectedSeqNum += 1;
-        assertEquals(txid2, resTxid);
-        assertEquals(table, resTableName);
-        assertEquals(projection, resProjection);
-
-
-        rs.next();
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        resSeqNum = rs.getInt(ProvenanceBuffer.PROV_QUERY_SEQNUM);
-        resTableName = rs.getString(ProvenanceBuffer.PROV_QUERY_TABLENAMES);
-        resProjection = rs.getString(ProvenanceBuffer.PROV_QUERY_PROJECTION);
-        assertEquals(expectedSeqNum, resSeqNum);
-        expectedSeqNum += 1;
-        assertEquals(txid2, resTxid);
-        assertEquals(table, resTableName);
-        assertEquals("*", resProjection);
-
-        rs.next();
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        resSeqNum = rs.getInt(ProvenanceBuffer.PROV_QUERY_SEQNUM);
-        resTableName = rs.getString(ProvenanceBuffer.PROV_QUERY_TABLENAMES);
-        resProjection = rs.getString(ProvenanceBuffer.PROV_QUERY_PROJECTION);
-        assertEquals(expectedSeqNum, resSeqNum);
-        expectedSeqNum += 1;
-        assertEquals(txid2, resTxid);
-        assertEquals(table, resTableName);
-        assertEquals("*", resProjection);
-
-        // Should be a read.
-        rs.next();
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        resSeqNum = rs.getInt(ProvenanceBuffer.PROV_QUERY_SEQNUM);
-        resTableName = rs.getString(ProvenanceBuffer.PROV_QUERY_TABLENAMES);
-        resProjection = rs.getString(ProvenanceBuffer.PROV_QUERY_PROJECTION);
-        assertEquals(expectedSeqNum, resSeqNum);
-        expectedSeqNum += 1;
-        assertEquals(txid2, resTxid);
-        assertEquals(table, resTableName);
-        assertEquals(projection, resProjection);
-
-        // Should be an update.
-        rs.next();
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        resSeqNum = rs.getInt(ProvenanceBuffer.PROV_QUERY_SEQNUM);
-        resTableName = rs.getString(ProvenanceBuffer.PROV_QUERY_TABLENAMES);
-        resProjection = rs.getString(ProvenanceBuffer.PROV_QUERY_PROJECTION);
-        assertEquals(expectedSeqNum, resSeqNum);
-        expectedSeqNum += 1;
-        assertEquals(txid2, resTxid);
-        assertEquals(table, resTableName);
-        assertEquals("*", resProjection);
-
-        // Should be a read again.
-        rs.next();
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        resSeqNum = rs.getInt(ProvenanceBuffer.PROV_QUERY_SEQNUM);
-        resTableName = rs.getString(ProvenanceBuffer.PROV_QUERY_TABLENAMES);
-        resProjection = rs.getString(ProvenanceBuffer.PROV_QUERY_PROJECTION);
-        assertEquals(expectedSeqNum, resSeqNum);
-        expectedSeqNum += 1;
-        assertEquals(txid2, resTxid);
-        assertEquals(table, resTableName);
-        assertEquals(projection, resProjection);
-
-        // Should be a delete.
-        rs.next();
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        resSeqNum = rs.getInt(ProvenanceBuffer.PROV_QUERY_SEQNUM);
-        resTableName = rs.getString(ProvenanceBuffer.PROV_QUERY_TABLENAMES);
-        resProjection = rs.getString(ProvenanceBuffer.PROV_QUERY_PROJECTION);
-        assertEquals(expectedSeqNum, resSeqNum);
-        expectedSeqNum += 1;
-        assertEquals(txid2, resTxid);
-        assertEquals(table, resTableName);
-        assertEquals("*", resProjection);
-    }
-
-    @Test
-    public void testPostgresProvenanceJoins() throws InvalidProtocolBufferException, SQLException, InterruptedException {
-        logger.info("testPostgresProvenanceJoins");
-
-        PostgresConnection conn = new PostgresConnection("localhost", ApiaryConfig.postgresPort, "postgres", "postgres", "dbos");
-
-        apiaryWorker = new ApiaryWorker(new ApiaryNaiveScheduler(), 4, ApiaryConfig.postgres, ApiaryConfig.provenanceDefaultAddress);
-        apiaryWorker.registerConnection(ApiaryConfig.postgres, conn);
-        apiaryWorker.registerFunction("PostgresProvenanceJoins", ApiaryConfig.postgres, PostgresProvenanceJoins::new);
-        apiaryWorker.startServing();
-
-        ProvenanceBuffer provBuff = apiaryWorker.workerContext.provBuff;
-        assert(provBuff != null);
-
-        Thread.sleep(ProvenanceBuffer.exportInterval * 4);
-        Connection provConn = provBuff.conn.get();
-        Statement stmt = provConn.createStatement();
-        String[] tables = {"FUNCINVOCATIONS", "KVTableEvents", "KVTableTwoEvents"};
-        for (String table : tables) {
-            stmt.execute(String.format("TRUNCATE TABLE %s;", table));
-        }
-
-        ApiaryWorkerClient client = new ApiaryWorkerClient("localhost");
-
-        int res;
-        res = client.executeFunction("PostgresProvenanceJoins", 1, 2, 3).getInt();
-        assertEquals(5, res);
-
-        Thread.sleep(ProvenanceBuffer.exportInterval * 2);
-
-        // Check KVTable.
-        String table = "KVTableEvents";
-        ResultSet rs = stmt.executeQuery(String.format("SELECT * FROM %s ORDER BY %s;", table, ProvenanceBuffer.PROV_APIARY_TIMESTAMP));
-        rs.next();
-
-        // Should be an insert for key=1.
-        int resExportOp = rs.getInt(ProvenanceBuffer.PROV_APIARY_OPERATION_TYPE);
-        int resKey = rs.getInt("KVKey");
-        int resValue = rs.getInt("KVValue");
-        assertEquals(ProvenanceBuffer.ExportOperation.INSERT.getValue(), resExportOp);
-        assertEquals(1, resKey);
-        assertEquals(2, resValue);
-
-        // Should be a read.
-        rs.next();
-        resExportOp = rs.getInt(ProvenanceBuffer.PROV_APIARY_OPERATION_TYPE);
-        resKey = rs.getInt("KVKey");
-        resValue = rs.getInt("KVValue");
-        assertEquals(ProvenanceBuffer.ExportOperation.READ.getValue(), resExportOp);
-        assertEquals(2, resValue);
-
-        // Check KVTable.
-        table = "KVTableTwoEvents";
-        rs = stmt.executeQuery(String.format("SELECT * FROM %s ORDER BY %s;", table, ProvenanceBuffer.PROV_APIARY_TIMESTAMP));
-        rs.next();
-
-        // Should be an insert for key=1.
-        resExportOp = rs.getInt(ProvenanceBuffer.PROV_APIARY_OPERATION_TYPE);
-        resKey = rs.getInt("KVKeyTwo");
-        resValue = rs.getInt("KVValueTwo");
-        assertEquals(ProvenanceBuffer.ExportOperation.INSERT.getValue(), resExportOp);
-        assertEquals(1, resKey);
-        assertEquals(3, resValue);
-
-        // Should be a read.
-        rs.next();
-        resExportOp = rs.getInt(ProvenanceBuffer.PROV_APIARY_OPERATION_TYPE);
-        resKey = rs.getInt("KVKeyTwo");
-        resValue = rs.getInt("KVValueTwo");
-        assertEquals(ProvenanceBuffer.ExportOperation.READ.getValue(), resExportOp);
-        assertEquals(3, resValue);
-    }
-
-    @Test
-    public void testPostgresProvenanceMultiRows() throws InvalidProtocolBufferException, SQLException, InterruptedException {
-        logger.info("testPostgresProvenanceMultiRows");
-
-        PostgresConnection conn = new PostgresConnection("localhost", ApiaryConfig.postgresPort, "postgres", "postgres", "dbos");
-
-        apiaryWorker = new ApiaryWorker(new ApiaryNaiveScheduler(), 4, ApiaryConfig.postgres, ApiaryConfig.provenanceDefaultAddress);
-        apiaryWorker.registerConnection(ApiaryConfig.postgres, conn);
-        apiaryWorker.registerFunction("PostgresProvenanceMultiRows", ApiaryConfig.postgres, PostgresProvenanceMultiRows::new);
-        apiaryWorker.startServing();
-
-        ProvenanceBuffer provBuff = apiaryWorker.workerContext.provBuff;
-        assert(provBuff != null);
-
-        // Wait a bit so previous provenance capture data would be flushed out.
-        Thread.sleep(ProvenanceBuffer.exportInterval * 4);
-        Connection provConn = provBuff.conn.get();
-        Statement stmt = provConn.createStatement();
-        String[] tables = {"FUNCINVOCATIONS", "KVTableEvents"};
-        for (String table : tables) {
-            stmt.execute(String.format("TRUNCATE TABLE %s;", table));
-        }
-
-        ApiaryWorkerClient client = new ApiaryWorkerClient("localhost");
-
-        int res;
-        int key1 = 10, value1 = 100;
-        int key2 = 20, value2 = 11;
-        res = client.executeFunction("PostgresProvenanceMultiRows", key1, value1, key2, value2).getInt();
-        assertEquals(111, res);
-
-        Thread.sleep(ProvenanceBuffer.exportInterval * 2);
-
-        // Check provenance tables.
-        // Check function invocation table.
-        String table = "FUNCINVOCATIONS";
-        ResultSet rs = stmt.executeQuery(String.format("SELECT * FROM %s ORDER BY %s DESC;", table, ProvenanceBuffer.PROV_APIARY_TIMESTAMP));
-        rs.next();
-        long txid1 = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        long resExecId = rs.getLong(ProvenanceBuffer.PROV_EXECUTIONID);
-        String resService = rs.getString(ProvenanceBuffer.PROV_SERVICE);
-        String resFuncName = rs.getString(ProvenanceBuffer.PROV_PROCEDURENAME);
-        assertEquals("DefaultService", resService);
-        assertEquals(PostgresProvenanceMultiRows.class.getName(), resFuncName);
-
-        // Check KVTable.
-        table = "KVTableEvents";
-        rs = stmt.executeQuery(String.format("SELECT * FROM %s ORDER BY %s, KVKEY;", table, ProvenanceBuffer.PROV_APIARY_TIMESTAMP));
-        rs.next();
-
-        // Should be an insert for key1.
-        long resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        int resExportOp = rs.getInt(ProvenanceBuffer.PROV_APIARY_OPERATION_TYPE);
-        int resKey = rs.getInt("KVKey");
-        int resValue = rs.getInt("KVValue");
-        assertEquals(txid1, resTxid);
-        assertEquals(ProvenanceBuffer.ExportOperation.INSERT.getValue(), resExportOp);
-        assertEquals(key1, resKey);
-        assertEquals(value1, resValue);
-
-        // Should be an insert for the key2.
-        rs.next();
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        assertEquals(txid1, resTxid);
-        resExportOp = rs.getInt(ProvenanceBuffer.PROV_APIARY_OPERATION_TYPE);
-        resKey = rs.getInt("KVKey");
-        resValue = rs.getInt("KVValue");
-        assertEquals(ProvenanceBuffer.ExportOperation.INSERT.getValue(), resExportOp);
-        assertEquals(key2, resKey);
-        assertEquals(value2, resValue);
-
-        // Should be a read for key1.
-        rs.next();
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        assertEquals(txid1, resTxid);
-        resExportOp = rs.getInt(ProvenanceBuffer.PROV_APIARY_OPERATION_TYPE);
-        resKey = rs.getInt("KVKey");
-        resValue = rs.getInt("KVValue");
-        assertEquals(ProvenanceBuffer.ExportOperation.READ.getValue(), resExportOp);
-        assertEquals(key1, resKey);
-        assertEquals(value1, resValue);
-
-        // Should be a read again for key2.
-        rs.next();
-        resTxid = rs.getLong(ProvenanceBuffer.PROV_APIARY_TRANSACTION_ID);
-        assertEquals(txid1, resTxid);
-        resExportOp = rs.getInt(ProvenanceBuffer.PROV_APIARY_OPERATION_TYPE);
-        resKey = rs.getInt("KVKey");
-        resValue = rs.getInt("KVValue");
-        assertEquals(ProvenanceBuffer.ExportOperation.READ.getValue(), resExportOp);
-        assertEquals(key2, resKey);
-        assertEquals(value2, resValue);
     }
 }
